add deploy_beaglebone.sh
sorry ccgx :-)

Usage: after bitbaking a new swu file, run this script to install it onto
a target. And after installing it also logs you in with ssh.

Beaglebone only since it uses the /scratch partition. Its obviously easy to
make it work on a CCGX as well by mandating an sdcard or something mounted
into the CCGX.

And, probably its not too hard either to fixup swupdate to accept a swu file
straight from stdin, and then all this temp stuff can be dropped.